backend/requirements.txt
fastapi
uvicorn
websockets
langgraph
langchain-google-genai
duckduckgo-search
langchain-community
pydantic
python-dotenv
